+++ packages/toolbars-v3/web/src/index.ts
@@ -19,4 +19,5 @@ export { pluginToolbarItemsRenders } from './pluginToolbarItemRenders';
 export * from './resolvers/tiptapResolvers';
 export { ToggleButton } from './components/buttons';
 export { default as RicosToolbarComponent } from './components/RicosToolbarComponent';
+export { default as FloatingAddPluginMenu } from './modals/FloatingPluginMenu/FloatingAddPluginMenu';
 export { withToolbarContext } from './utils/withContext';