+++ packages/toolbars-v3/web/src/modals/FloatingPluginMenu/EditorSelectionToPosition.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import { posToDOMRect } from '../utils/posToDOMRect';
+import { posToDOMRect } from '@tiptap/core';
 import { debounce } from 'lodash';
 
 const EditorSelectionToPosition = ({ children, editor }) => {