+++ packages/toolbars-v3/web/src/modals/FloatingPluginMenu/PlusButton.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { CirclePlusIcon as PlusIcon } from 'wix-rich-content-ui-components';
-import styles from '../../../statics/styles/floating-add-plugin-menu.scss';
+import styles from './styles/floating-add-plugin-menu.scss';
 
 type PlusButtonProps = {
   onClick: () => void;