+++ packages/toolbars-v3/web/src/modals/FloatingPluginMenu/types.ts
@@ -1,0 +1,5 @@
+import type { AddButton } from 'wix-rich-content-common';
+
+export interface IPluginMenuButtonClick {
+  (modal: AddButton['modal'], command: AddButton['command']): void;
+}